requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
cloudpickle==0.5.3
cycler==0.10.0
dask==0.18.2
decorator==4.3.0
docopt==0.6.1
docopts==0.6.1
kiwisolver==1.0.1
matplotlib==2.2.2
more-itertools==4.2.0
networkx==2.1
numpy==1.14.5
opencv-python==3.4.2.17
Pillow==5.2.0
pluggy==0.6.0
py==1.5.4
pycparser==2.18
pyparsing==2.2.0
pytest==3.6.3
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
sortedcontainers==2.0.4
tifffile==0.15.1
toolz==0.9.0
